requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.7.2
joblib==1.4.2
rapidfuzz==3.10.0
python-dateutil==2.9.0
pyodbc==5.1.0
scipy==1.13.0
matplotlib==3.9.0
seaborn==0.13.1
streamlit==1.32.0
plotly==5.18.0
